stage 2

+++ app/src/main/java/com/yae/torrenthelper/data/models.kt
@@ -1,6 +1,6 @@
 package com.yae.torrenthelper.data
 
-data class TarInfo(val path:String, val size:Int, val url:String)
+data class TarInfo(val path:String, val size:Long, val url:String)
 
 data class TorrentInfo(
     val name:String,
